增加unilm预训练模式（试验）

+++ pretraining/data_utils.py
@@ -316,7 +316,7 @@ class TrainingDatasetUniLM(TrainingDatasetGPT):
             segment_ids = K.one_hot(segment, sequence_length)
             segment_ids = K.cast(K.cumsum(segment_ids), 'int64')
             token_ids_1 = token_ids[:segment]
-            token_ids_2 = K.zeros_like(token_ids[:1]) + token_sep_id
+            token_ids_2 = K.zeros_like(token_ids[:1], dtype='int64') + token_sep_id
             token_ids_3 = token_ids[segment:-1]
             token_ids = K.concatenate([token_ids_1, token_ids_2, token_ids_3])
             x = {
